cleanup shutdown function
